+++ UDPSocket.h
@@ -29,20 +29,20 @@
 
 class CUDPSocket {
 public:
-	CUDPSocket(const std::string& address, unsigned int port = 0U);
-	CUDPSocket(unsigned int port = 0U);
+	CUDPSocket(const std::string& address, unsigned short port = 0U);
+	CUDPSocket(unsigned short port = 0U);
 	~CUDPSocket();
 
 	bool open(unsigned int af = AF_UNSPEC);
-	bool open(const unsigned int af, const std::string& address, const unsigned int port);
+	bool open(const unsigned int af, const std::string& address, const unsigned short port);
 
 	int  read(unsigned char* buffer, unsigned int length, sockaddr_storage& address, unsigned int &address_length);
 	bool write(const unsigned char* buffer, unsigned int length, const sockaddr_storage& address, unsigned int address_length);
 
 	void close();
 
-	static int lookup(const std::string& hostName, unsigned int port, sockaddr_storage& address, unsigned int& address_length);
-	static int lookup(const std::string& hostName, unsigned int port, sockaddr_storage& address, unsigned int& address_length, struct addrinfo& hints);
+	static int lookup(const std::string& hostName, unsigned short port, sockaddr_storage& address, unsigned int& address_length);
+	static int lookup(const std::string& hostName, unsigned short port, sockaddr_storage& address, unsigned int& address_length, struct addrinfo& hints);
 
 	static bool match(const sockaddr_storage& addr1, const sockaddr_storage& addr2);
